refactor: rename transaction modal workflow folder
