+++ frontend/src/transactions/TransactionsPage.tsx
@@ -10,7 +10,7 @@ import {
 } from '@/api/transactions'
 import { transactionKeys } from '@/api/cache/queryKeys'
 import TransactionListSection from '@/transactions/components/TransactionListSection'
-import CreateTransactionModal from '@/transactions/components/transaction-modal/CreateTransactionModal'
+import CreateTransactionModal from '@/transactions/components/TransactionModal/CreateTransactionModal'
 import TransactionsTopBand from '@/transactions/components/TransactionsTopBand'
 import type { TransactionListFilters } from '@/transactions/types/transactionList'
 import {
