+++ frontend/src/transactions/components/TransactionModal/TransactionModalFieldLabelRow.tsx
@@ -1,6 +1,6 @@
 import type { ReactNode } from 'react'
 import { AnimatePresence, motion } from 'motion/react'
-import { EASE } from '@/transactions/components/transaction-modal/transactionModalConstants'
+import { EASE } from '@/transactions/components/TransactionModal/transactionModalConstants'
 import { joinClassNames } from '@/utils/classNames'
 
 interface TransactionModalFieldLabelRowProps {
