+++ frontend/src/transactions/components/TransactionModal/TransactionModalSectionFrame.tsx
