+++ frontend/src/transactions/components/TransactionModal/hooks/useTransactionModalEnvironment.ts
