+++ frontend/tests/transactions/transactionModal.test.ts
@@ -6,19 +6,19 @@ import type { AccountsOverview } from '@/api/accounts'
 import type { Category } from '@/api/categories'
 import type { Currency } from '@/api/currency'
 import type { Transaction } from '@/api/transactions'
-import { buildCategoryOptions } from '@/transactions/components/transaction-modal/transactionModalCategories'
-import { buildInitialTransactionForm } from '@/transactions/components/transaction-modal/transactionModalInitialForm'
+import { buildCategoryOptions } from '@/transactions/components/TransactionModal/transactionModalCategories'
+import { buildInitialTransactionForm } from '@/transactions/components/TransactionModal/transactionModalInitialForm'
 import {
   amountInputToMinorUnits,
   amountToInputString,
   applyTransactionDirection,
   getDirectionFromAmountInputSign,
-} from '@/transactions/components/transaction-modal/transactionModalMoney'
+} from '@/transactions/components/TransactionModal/transactionModalMoney'
 import {
   buildCreateTransactionPayload,
   buildUpdateTransactionPatch,
-} from '@/transactions/components/transaction-modal/transactionModalPayloads'
-import { validateTransactionForm } from '@/transactions/components/transaction-modal/transactionModalValidation'
+} from '@/transactions/components/TransactionModal/transactionModalPayloads'
+import { validateTransactionForm } from '@/transactions/components/TransactionModal/transactionModalValidation'
 
 const currencies: Currency[] = [
   { id: 'CAD', name: 'Canadian Dollar', symbol: '$', minor_unit_exponent: 2 },
